Uppercase words starting with a digit
In general, first character of each word can be uppercased to get a
proper camel case when joined. But, sometimes those words can start with
a digit which likely mark the beginning of an acronym such as "3D", in
such cases they better be uppercased completely (e.g. QVector3D).