Config for namespaces to filter out during generation
Some libraries might have namespaces which map their library name. In
this case the generator would have the rust types end up in a module
named "lib::lib" which feels redundant. In such a case, we now allow the
generator to ignore the namespace of the type to generate the rust name,
allowing to simplify to just "lib" for the module name.