Add an example for Qt 3D use from Rust
It's not really pleasant to use yet due to the amount of unsafe methods
and the into_raw calls which are needed to leave the life time
management to QObject's parent/child tree.

Clearly comes from QObject's parent/child tree fighting with Rust own
memory management and Qt 3D using heavily composition and not inheritance.
My goal is now to change the generator to simplify this code.